The executable documentation runs where its steps do, and says when it does not
A spec case's steps are POSIX shell run through bash. Windows has no such
shell, so on the new Windows CI leg the documentation suite could only ever
fail — 158 cases red for a reason that says nothing about the tool.

It skips there now, with a reason naming issue #137, so the gap is a line in
the summary rather than an absence. The parser tests still run everywhere: the
skip sits in the helper that reaches a step, not on the module, and nothing
that only parses needs a shell.

Splitting test_harness.py fell out of it — the file was one line under the
threshold and its own section comments had been naming three concerns for a
while. Markers and their normalisation are what one step means; contexts are
the rules between cases; the runner both drive is neither.
