+++ tests/harness/__init__.py
@@ -8,6 +8,16 @@ contract is ``docs/executable_spec.md``. pytest is the runner (see
 from .errors import SpecError, SpecFailure
 from .parser import SpecCase, parse_spec
 from .runner import execute
+from .steps import POSIX_SHELL_ONLY, STEPS_RUN_ON_THIS_PLATFORM
 from .text import normalize
 
-__all__ = ["SpecCase", "SpecError", "SpecFailure", "execute", "normalize", "parse_spec"]
+__all__ = [
+    "POSIX_SHELL_ONLY",
+    "STEPS_RUN_ON_THIS_PLATFORM",
+    "SpecCase",
+    "SpecError",
+    "SpecFailure",
+    "execute",
+    "normalize",
+    "parse_spec",
+]
