+++ tests/test_spec_markers.py
@@ -1,29 +1,14 @@
-"""Unit tests for the spec harness — one per marker plus the cross-cutting rules."""
+"""What each marker in the spec grammar means, one marker at a time.
 
-from pathlib import Path
-
-import pytest
-
-from harness import SpecError, SpecFailure, execute, normalize, parse_spec
-
-
-def _status(test, where: Path, repo_root: Path) -> str:
-    """Run one parsed test in its own dir; return "skip"/"pass"/"fail"."""
-    if test.skip:
-        return "skip"
-    where.mkdir()
-    try:
-        execute(test, where, repo_root)
-        return "pass"
-    except (SpecFailure, SpecError):
-        return "fail"
+The grammar contract is ``docs/executable_spec.md``. Normalisation sits here
+too: it is how a screen marker decides two outputs are the same, so it is part
+of what a marker means rather than a separate concern.
+"""
 
 
-def run(text: str, tmp_path: Path, repo_root: Path | None = None) -> list[str]:
-    """Parse + run a spec body, returning a status per test."""
-    root = repo_root or tmp_path
-    cases = parse_spec(text)
-    return [_status(c, tmp_path / f"t{i}", root) for i, c in enumerate(cases)]
+from harness import normalize, parse_spec
+
+from spec_runs import run
 
 
 # --- normalisation ---------------------------------------------------------
@@ -134,54 +119,3 @@ def test_bash_block_never_consumed_as_payload(tmp_path):
 def test_bare_block_without_marker_is_ignored(tmp_path):
     spec = "# T\n```toml\njust docs\n```\n```bash\ntrue\n```\n"
     assert run(spec, tmp_path) == ["pass"]
-
-
-# --- contexts --------------------------------------------------------------
-
-
-def test_sibling_contexts_are_isolated(tmp_path):
-    spec = (
-        "# Root\n"
-        "## A\n```bash\necho hi > shared.txt\n```\n"
-        "## B\n```bash\ncat shared.txt\n```\n"
-    )
-    # B runs in a fresh dir, so shared.txt is absent and B fails.
-    assert run(spec, tmp_path) == ["pass", "fail"]
-
-
-def test_ancestor_preamble_accumulates(tmp_path):
-    spec = (
-        "# Root\n✏️A\n```text\n1\n```\n"
-        "## Mid\n✏️B\n```text\n2\n```\n"
-        "### Leaf\n```bash\nprintf '%s%s' \"$A\" \"$B\"\n```\n🖥️ ✅\n```text\n12\n```\n"
-    )
-    assert run(spec, tmp_path) == ["pass"]
-
-
-def test_only_leaves_are_tests(tmp_path):
-    spec = "# Root\n## A\n```bash\ntrue\n```\n## B\n```bash\ntrue\n```\n"
-    assert run(spec, tmp_path) == ["pass", "pass"]
-
-
-def test_skip_is_reported_not_run(tmp_path):
-    spec = "# T 🟡\n```bash\nexit 1\n```\n"
-    results = parse_spec(spec)
-    assert len(results) == 1 and results[0].skip is True
-    assert run(spec, tmp_path) == ["skip"]
-
-
-def test_skip_inherited_from_ancestor(tmp_path):
-    spec = "# Group 🟡\n## Leaf\n```bash\nexit 1\n```\n"
-    assert run(spec, tmp_path) == ["skip"]
-
-
-def test_missing_required_block_is_spec_error():
-    # ✏️ with no following block is malformed (caught while pairing markers).
-    with pytest.raises(SpecError):
-        parse_spec("# T\n✏️X\n```bash\ntrue\n```\n")
-
-
-def test_stdin_missing_block_is_spec_error():
-    # ⌨️ likewise requires its payload block before the command runs.
-    with pytest.raises(SpecError):
-        parse_spec("# T\n⌨️\n```bash\ncat\n```\n")
